+++ recon/symbolized/app/FUN_000232b0.c
@@ -1,8 +1,12 @@
 #include "g1_app_symbols.h"
-/* named: FUN_000232b0 */
+/* readable reconstruction; identity: FUN_000232b0 @ 0x000232b0
+ * public-name: FUN_000232b0
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_device_serial_buf                      @ 0x20003003
+ */
 /* Reconstructed FUN_000232b0 @ 0x232b0  (parity: 300/300 trials, PROVEN) */
 
 unsigned int FUN_000232b0(void){
-  return ((uintptr_t)&g_device_serial_buf) /*=0x20003003*/;
+  return ((unsigned long)&g_device_serial_buf) /*=0x20003003*/;
 }
-
