+++ recon/symbolized/app/FUN_000232b8.c
@@ -1,11 +1,12 @@
 #include "g1_app_symbols.h"
-/* named: FUN_000232b8 */
-/* globals referenced:
-//   0x20002fe3  g_product_code_buf           
-*/
+/* readable reconstruction; identity: FUN_000232b8 @ 0x000232b8
+ * public-name: FUN_000232b8
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_product_code_buf                       @ 0x20002fe3
+ */
 /* Reconstructed FUN_000232b8 @ 0x232b8  (parity: 300/300 trials, PROVEN) */
 
 unsigned int FUN_000232b8(void){
-  return ((uintptr_t)&g_product_code_buf) /*=0x20002fe3*/;
+  return ((unsigned long)&g_product_code_buf) /*=0x20002fe3*/;
 }
-
