+++ recon/symbolized/app/FUN_00023ee0.c
@@ -1,11 +1,12 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00023ee0 */
-/* globals referenced:
-//   0x20018462  g_ui_mode_flag               
-*/
+/* readable reconstruction; identity: FUN_00023ee0 @ 0x00023ee0
+ * public-name: FUN_00023ee0
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_ui_mode_flag                           @ 0x20018462
+ */
 /* Reconstructed FUN_00023ee0 @ 0x23ee0  (parity: 300/300 trials, PROVEN) */
 
 unsigned char FUN_00023ee0(void) {
-    return *(volatile unsigned char*)(((uintptr_t)&g_ui_mode_flag) /*=0x20018462*/ + 1);
+    return *(volatile unsigned char*)(((unsigned long)&g_ui_mode_flag) /*=0x20018462*/ + 1);
 }
-
