+++ recon/symbolized/app/FUN_00023eec.c
@@ -1,11 +1,12 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00023eec */
-/* globals referenced:
-//   0x20018462  g_ui_mode_flag               
-*/
+/* readable reconstruction; identity: FUN_00023eec @ 0x00023eec
+ * public-name: FUN_00023eec
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_ui_mode_flag                           @ 0x20018462
+ */
 /* Reconstructed FUN_00023eec @ 0x23eec  (parity: 300/300 trials, PROVEN) */
 
 unsigned char FUN_00023eec(void) {
-    return *(volatile unsigned char*)(((uintptr_t)&g_ui_mode_flag) /*=0x20018462*/ + 2);
+    return *(volatile unsigned char*)(((unsigned long)&g_ui_mode_flag) /*=0x20018462*/ + 2);
 }
-
