+++ recon/symbolized/app/FUN_00024d50.c
@@ -1,8 +1,10 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00024d50 */
-/* globals referenced:
-//   0x20007a44  g_st25dv_dev                 
-*/
+/* readable reconstruction; identity: FUN_00024d50 @ 0x00024d50
+ * public-name: FUN_00024d50
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_st25dv_dev                             @ 0x20007a44
+ */
 /* Reconstructed FUN_00024d50 @ 0x24d50  (parity: 300/300 trials, PROVEN) */
 
 extern void FUN_0007c968(unsigned int, void *, unsigned int, unsigned int, unsigned int);
@@ -15,8 +17,7 @@ int FUN_00024d50(unsigned int param_1, unsigned int param_2, unsigned int param_
 
   local_c = param_2;
   uStack_8 = param_3;
-  FUN_0007c968(*(volatile unsigned int*)((uintptr_t)&g_st25dv_dev) /*=0x20007a44*/,&local_c,param_3,((uintptr_t)&g_st25dv_dev) /*=0x20007a44*/,param_1);
+  FUN_0007c968(*(volatile unsigned int*)((unsigned long)&g_st25dv_dev) /*=0x20007a44*/,&local_c,param_3,((unsigned long)&g_st25dv_dev) /*=0x20007a44*/,param_1);
   iVar1 = (int)(local_c >> 0x10) + 1;
   return (int)(local_c & 0xff) * iVar1 + iVar1;
 }
-
