+++ recon/symbolized/app/FUN_00030c24.c
@@ -1,5 +1,10 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00030c24 */
+/* readable reconstruction; identity: FUN_00030c24 @ 0x00030c24
+ * public-name: FUN_00030c24
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_misc_dev_handle_2418                   @ 0x20002418
+ */
 /* Reconstructed FUN_00030c24 @ 0x30c24  (parity: 300/300 trials, PROVEN) */
 
 typedef int (*fnptr)(unsigned int, void *, int, int);
@@ -22,11 +27,10 @@ int FUN_00030c24(unsigned int param_1, unsigned int param_2, unsigned int param_
   uStack_1c = param_2;
   local_14 = param_3;
   uStack_10 = param_4;
-  base = *(volatile unsigned int *)((uintptr_t)&g_misc_dev_handle_2418) /*=0x20002418*/;
+  base = *(volatile unsigned int *)((unsigned long)&g_misc_dev_handle_2418) /*=0x20002418*/;
   field = *(int *)(base + 8);
   fp = *(fnptr *)(field + 8);
   iVar1 = fp(base, &local_20, 2, 0x12);
   if (iVar1 != 0) iVar1 = 1;
   return -iVar1;
 }
-
