+++ recon/symbolized/app/FUN_00033088.c
@@ -1,20 +1,21 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00033088 */
-/* globals referenced:
-//   0x20002424  g_fuel_gauge_dev             
-*/
+/* readable reconstruction; identity: FUN_00033088 @ 0x00033088
+ * public-name: FUN_00033088
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_fuel_gauge_dev                         @ 0x20002424
+ */
 /* Reconstructed FUN_00033088 @ 0x33088  (parity: 300/300 trials, PROVEN) */
 
 int FUN_00033088(int param_1)
 {
-  int iVar1 = ((uintptr_t)&g_fuel_gauge_dev) /*=0x20002424*/;
+  int iVar1 = ((unsigned long)&g_fuel_gauge_dev) /*=0x20002424*/;
   if (param_1 != 0) {
-    if (*(volatile int *)(((uintptr_t)&g_fuel_gauge_dev) /*=0x20002424*/ + 4) != param_1) {
-      *(volatile int *)(((uintptr_t)&g_fuel_gauge_dev) /*=0x20002424*/ + 4) = param_1;
+    if (*(volatile int *)(((unsigned long)&g_fuel_gauge_dev) /*=0x20002424*/ + 4) != param_1) {
+      *(volatile int *)(((unsigned long)&g_fuel_gauge_dev) /*=0x20002424*/ + 4) = param_1;
     }
     *(volatile unsigned char *)(iVar1 + 8) = 1;
     return 0;
   }
   return -1;
 }
-
