+++ recon/symbolized/app/FUN_00034390.c
@@ -1,11 +1,12 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00034390 */
-/* globals referenced:
-//   0x2001a22a  g_message_pool_index         
-*/
+/* readable reconstruction; identity: FUN_00034390 @ 0x00034390
+ * public-name: FUN_00034390
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_message_pool_index                     @ 0x2001a22a
+ */
 /* Reconstructed FUN_00034390 @ 0x34390  (parity: 300/300 trials, PROVEN) */
 
 unsigned char FUN_00034390(void) {
-    return *(volatile unsigned char*)((uintptr_t)&g_message_pool_index) /*=0x2001a22a*/;
+    return *(volatile unsigned char*)((unsigned long)&g_message_pool_index) /*=0x2001a22a*/;
 }
-
