+++ recon/symbolized/app/FUN_0003dfe4.c
@@ -1,13 +1,16 @@
 #include "g1_app_symbols.h"
-/* named: FUN_0003dfe4 */
-/* globals referenced:
-//   0x20004bb8  g_ipc_or_evt_ctx             
-*/
+/* readable reconstruction; identity: FUN_0003dfe4 @ 0x0003dfe4
+ * public-name: FUN_0003dfe4
+ * durable-map: recon/catalogs/function_names_app.json
+ * callees (readable <= raw @ address):
+ *   memset_bytes                             <= FUN_00086c78 @ 0x00086c78
+ * address symbols (name @ address):
+ *   g_ipc_or_evt_ctx                         @ 0x20004bb8
+ */
 /* Reconstructed FUN_0003dfe4 @ 0x3dfe4  (parity: 300/300 trials, PROVEN) */
 
 extern void memset_bytes(unsigned int a, unsigned int b, unsigned int c);
 unsigned int FUN_0003dfe4(void) {
-  memset_bytes((g_ipc_or_evt_ctx) /*=0x20004bb8*/, 0, 0x38);
+  memset_bytes(((unsigned long)g_ipc_or_evt_ctx) /*=0x20004bb8*/, 0, 0x38);
   return 0;
 }
-
