+++ recon/symbolized/app/FUN_000429f8.c
@@ -1,9 +1,13 @@
 #include "g1_app_symbols.h"
-/* named: FUN_000429f8 */
-/* globals referenced:
-//   0x2001cdce  onboarding_secondary_reset_flag 
-//   0x2001cdd1  onboarding_retry_counter     
-*/
+/* readable reconstruction; identity: FUN_000429f8 @ 0x000429f8
+ * public-name: FUN_000429f8
+ * durable-map: recon/catalogs/function_names_app.json
+ * callees (readable <= raw @ address):
+ *   get_device_info                          <= FUN_000167a8 @ 0x000167a8
+ * address symbols (name @ address):
+ *   onboarding_secondary_reset_flag          @ 0x2001cdce
+ *   onboarding_retry_counter                 @ 0x2001cdd1
+ */
 /* Reconstructed FUN_000429f8 @ 0x429f8  (parity: 300/300 trials, PROVEN) */
 
 #include <stdint.h>
@@ -15,10 +19,9 @@ void FUN_000429f8(void){
        (iVar1=get_device_info(), *(char*)(*(int*)(iVar1+0x1014)+2)==0x13)) ||
       (iVar1=get_device_info(), *(char*)(*(int*)(iVar1+0x1014)+2)==0x14)) ||
      (iVar1=get_device_info(), *(char*)(*(int*)(iVar1+0x1014)+2)==0x15)) &&
-    ((*(volatile char*)((uintptr_t)&onboarding_secondary_reset_flag) /*=0x2001cdce*/==2) &&
+    ((*(volatile char*)((unsigned long)&onboarding_secondary_reset_flag) /*=0x2001cdce*/==2) &&
      (iVar1=get_device_info(), *(char*)(*(int*)(iVar1+0x1014)+3)==1))){
-    *(volatile char*)((uintptr_t)&onboarding_retry_counter) /*=0x2001cdd1*/=1;
+    *(volatile char*)((unsigned long)&onboarding_retry_counter) /*=0x2001cdd1*/=1;
   }
   return;
 }
-
