+++ recon/symbolized/app/FUN_00048b44.c
@@ -1,5 +1,10 @@
 #include "g1_app_symbols.h"
-/* named: FUN_00048b44 */
+/* readable reconstruction; identity: FUN_00048b44 @ 0x00048b44
+ * public-name: FUN_00048b44
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_dashboard_display_level                @ 0x20002544
+ */
 /* Reconstructed FUN_00048b44 @ 0x48b44  (parity: 300/300 trials, PROVEN) */
 
 unsigned int FUN_00048b44(int param_1)
@@ -10,7 +15,6 @@ unsigned int FUN_00048b44(int param_1)
     if (param_1 < 0x21) {
         param_1 = 0x21;
     }
-    *(volatile unsigned int*)((uintptr_t)&g_dashboard_display_level) /*=0x20002544*/ = param_1;
+    *(volatile unsigned int*)((unsigned long)&g_dashboard_display_level) /*=0x20002544*/ = param_1;
     return 0;
 }
-
