+++ recon/symbolized/app/FUN_0004b1c0.c
@@ -1,9 +1,13 @@
 #include "g1_app_symbols.h"
-/* named: FUN_0004b1c0 */
+/* readable reconstruction; identity: FUN_0004b1c0 @ 0x0004b1c0
+ * public-name: FUN_0004b1c0
+ * durable-map: recon/catalogs/function_names_app.json
+ * address symbols (name @ address):
+ *   g_misc_val_27c8                          @ 0x200027c8
+ */
 /* Reconstructed FUN_0004b1c0 @ 0x4b1c0  (parity: 300/300 trials, PROVEN) */
 
 #include <stdint.h>
 void FUN_0004b1c0(uint32_t param_1) {
-    *(volatile uint32_t*)((uintptr_t)&g_misc_val_27c8) /*=0x200027c8*/ = param_1;
+    *(volatile uint32_t*)((unsigned long)&g_misc_val_27c8) /*=0x200027c8*/ = param_1;
 }
-
